move files to flat directory
