+++ packages/mesh/src/base.ts
@@ -1,4 +1,4 @@
-import { DuplexStream } from "../utils";
+import { DuplexStream } from "./duplex-stream";
 
 interface DuplexIterableIterator<TInput, TOutput> extends AsyncIterableIterator<TOutput> {
   next(value?: TInput): Promise<IteratorResult<TOutput>>;
