+++ packages/mesh/src/channel.ts
@@ -1,17 +1,14 @@
-import { createRemoteDispatcher } from "./remote";
+import { pump } from "./pump";
+import { remote } from "./remote";
+import { createDuplexStream } from "./duplex-stream";
 import { StreamableDispatcher, Dispatcher } from "./base";
-import { 
-  pump,
-  DuplexStream,
-  createDuplexStream,
-} from "../utils";
 
 /**
  * Creates a new messaging channel over an existing message stream.
  */
 
 export const createChannelDispatcher = <TMessage>(channel: AsyncIterableIterator<any>, localDispatch: Dispatcher<any, any> = (() => {}), info?: any) => {
-   const remoteDispatch = createRemoteDispatcher({
+   const remoteDispatch = remote({
     info: info,
     adapter: {
       send(message) {
