+++ packages/mesh/src/conditional.ts
@@ -1,0 +1,10 @@
+import { Dispatcher } from "./base";
+import { wrapAsyncIterableIterator } from "./wrap-async-iterable-iterator";
+
+const noop = () => {};
+
+/**
+ * Executes a message against target dispatcher if filter returns true, otherwise execute a message against falsy dispatcher is provided.
+ */
+
+export const conditional = <T>(_if: (message: T) => boolean, _then: Dispatcher<T, any> = noop, _else: Dispatcher<T, any> = noop) => (message: T) => wrapAsyncIterableIterator((_if(message) ? _then : _else)(message))
