--- packages/mesh/src/dispatchers/index.ts
@@ -1,10 +1,0 @@
-export * from "./base";
-export * from "./fanout";
-export * from "./remote";
-export * from "./conditional";
-export * from "./proxy";
-export * from "./channel"
-export * from "./parallel";
-export * from "./sequence";
-export * from "./random";
-export * from "./round-robin";
