+++ packages/mesh/src/index.ts
@@ -1,3 +1,22 @@
-export * from "./dispatchers";
-export * from "./utils";
-export * from "./messages";
+import "reflect-metadata";
+(Symbol as any).asyncIterator = Symbol.asyncIterator || Symbol("Symbol.asyncIterator");
+
+export * from "./base";
+export * from "./combine";
+export * from "./remote";
+export * from "./conditional";
+export * from "./proxy";
+export * from "./channel"
+export * from "./parallel";
+export * from "./sequence";
+export * from "./fallback";
+export * from "./random";
+export * from "./round-robin";
+export * from "./through";
+export * from "./pipe";
+export * from "./pump";
+export * from "./deferred-promise";
+export * from "./queue";
+export * from "./wrap-async-iterable-iterator";
+export * from "./duplex-stream";
+export * from "./read-all";
