--- packages/mesh/src/messages/index.ts
@@ -1,3 +1,0 @@
-export interface Message {
-  readonly type: string;
-}
