+++ packages/mesh/src/parallel-test.ts
@@ -1,16 +1,16 @@
 import { expect } from "chai";
-import { createParallelDispatcher, readAll } from "..";
+import { parallel, readAll } from ".";
 
 describe(__filename + "#", () => {
   it("can be created", () => {
-    createParallelDispatcher([]);
+    parallel();
   });
 
   it("dispatch a message against one entry", async () => {
     let i = 0;
-    const dispatch = createParallelDispatcher([
+    const dispatch = parallel(
       m => i++
-    ]);
+    );
 
     dispatch({}).next();
     dispatch({}).next();
@@ -20,11 +20,11 @@ describe(__filename + "#", () => {
 
   it("dispatches messages in parallel", async () => {
     let i = 0;
-    const dispatch = createParallelDispatcher([
+    const dispatch = parallel(
       m => i++,
       m => i++,
       m => i++
-    ]);
+    );
 
     dispatch({}).next();
     expect(i).to.eql(3);
@@ -32,21 +32,21 @@ describe(__filename + "#", () => {
 
   it("dispatches a message against multiple dispatchers", async () => {
     let i = 0;
-    const dispatch = createParallelDispatcher([
+    const dispatch = parallel(
       m => i++,
       m => i++,
       m => i++
-    ]);
+    );
 
     expect(await readAll(dispatch({}))).to.eql([0, 1, 2]);
   });
 
   it("Can handle a dispatch that returns a rejection", async () => {
     let i = 0;
-    const dispatch = createParallelDispatcher([
+    const dispatch = parallel(
       m => i++,
       m => Promise.reject(new Error("some error"))
-    ]);
+    );
 
     let error;
     try {
@@ -60,17 +60,27 @@ describe(__filename + "#", () => {
   });
 
 
-  it("can nest parallel dispatchses", async () => {
+  it("can nest parallel dispatches", async () => {
     let i = 0;
-    const dispatch = createParallelDispatcher([
-      createParallelDispatcher([
+    const dispatch = parallel(
+      parallel(
         m => i++,
         m => i++,
         m => i++
-      ]),
+      ),
       m => i++
-    ]);
+    );
 
     expect(await readAll(dispatch({}))).to.eql([0, 1, 2, 3]);
+  });
+
+  it("can pass multiple arguments", async () => {
+    let i = 0;
+    const add = parallel(
+      (a, b) => a + b,
+      (a, b) => a + b
+    ) as (a: number, b: number) => AsyncIterableIterator<number>;
+
+    expect(await readAll(add(1, 2))).to.eql([3, 3]);
   });
 });
