+++ packages/mesh/src/round-robin.ts
@@ -1,11 +1,11 @@
-import { FanoutDispatcherTargetsParamType, createFanoutDispatcher} from "./fanout";
+import { combine, FanoutDispatcherTargetsParamType } from "./combine";
 
 /**
  * Executes a message against one target dispatcher that is rotated with each message.
  */
 
 
-export const createRoundRobinDispatcher = <T>(targets: FanoutDispatcherTargetsParamType<T>) => createFanoutDispatcher(targets, (<T>() => {
+export const roundRobin = <T>(...fns: Function[]) => combine(fns, (<T>() => {
   let current = 0;
   return (items: T[], each: (value: T) => any) => {
     let prev = current;
