+++ packages/mesh/src/sequence-test.ts
@@ -1,16 +1,16 @@
 import { expect } from "chai";
-import { timeout } from "../test";
-import { createSequenceDispatcher, pipe, readAll, through } from "..";
+import { timeout } from "./test";
+import { sequence, pipe, readAll, through } from ".";
 
 describe(__filename + "#", () => {
   it("can dispatch a message to multiple endpoints in sequence", async () => {
     let i = 0;
 
-    const dispatch = createSequenceDispatcher([
+    const dispatch = sequence(
       m => i++,
       m => i++,
       m => i++
-    ]);
+    );
 
     const iterable = dispatch({});
     expect((await iterable.next()).value).to.eql(0);
@@ -21,7 +21,7 @@ describe(__filename + "#", () => {
   it("can yield all values from targets", async () => {
     let i = 0;
 
-    const dispatch = createSequenceDispatcher([
+    const dispatch = sequence(
       function *(m) {
         yield "a";
         yield "b";
@@ -34,7 +34,7 @@ describe(__filename + "#", () => {
         yield "e";
         yield "f";
       }
-    ]);
+    );
 
     const iterable = dispatch({});
     expect((await iterable.next()).value).to.eql("a");
@@ -47,36 +47,24 @@ describe(__filename + "#", () => {
   });
 
   it("ignores returned values from target dispatchers", async () => {
-    const dispatch = createSequenceDispatcher([
+    const dispatch = sequence(
       function *(m) {
         yield "a";
         yield "b";
         return "c";
       }
-    ]);
+    );
     const iterable = dispatch({});
     expect((await iterable.next()).value).to.eql("a");
     expect((await iterable.next()).value).to.eql("b");
     expect((await iterable.next()).value).to.eql(undefined);
   });
 
-  it("can use a function that returns an array of target dispatchers", async () => {
-    const dispatch = createSequenceDispatcher(() => [
-      function*() { yield "a"; },
-      function*() { yield "b"; }
-    ]);
-
-    const iterable = dispatch({});
-    expect((await iterable.next()).value).to.eql("a");
-    expect((await iterable.next()).value).to.eql("b");
-    expect((await iterable.next()).done).to.eql(true);
-  });
-
   it("can write data to all dispatchers", async () => {
-    const dispatch = createSequenceDispatcher(() => [
+    const dispatch = sequence(
       m => through((chunk: string) => chunk.toUpperCase()),
       m => through((chunk: string) => `${chunk}!`)
-    ]);
+    );
 
     const iterable = dispatch({});
     expect((await iterable.next("a")).value).to.eql("A");
@@ -85,10 +73,10 @@ describe(__filename + "#", () => {
   });
 
   it("can write data to all dispatchers", async () => {
-    const dispatch = createSequenceDispatcher(() => [
+    const dispatch = sequence(
       m => through((chunk: string) => chunk.toUpperCase()),
       m => through((chunk: string) => `${chunk}!`)
-    ]);
+    );
 
     const iterable = dispatch({});
     expect((await iterable.next("a")).value).to.eql("A");
@@ -97,28 +85,28 @@ describe(__filename + "#", () => {
   });
 
   it("can pipe input and receive output", async () => {
-    const dispatch = createSequenceDispatcher(() => [
+    const dispatch = sequence(
       m => through((v: number) => -v)
-    ]);
+    );
 
     expect(await readAll(pipe([1, 2, 3], dispatch({})))).to.eql([-1, -2, -3]);
   });
 
   it("can nest sequence dispatchers", async () => {
     let i = '';
-    const dispatch = createSequenceDispatcher([
-      createSequenceDispatcher([
+    const dispatch = sequence(
+      sequence(
         m => i += 'a',
         m => i += 'b',
         m => i += 'c'
-      ]),
-      createSequenceDispatcher([
+      ),
+      sequence(
         m => i += 'd',
         m => i += 'e',
         m => i += 'f'
-      ]),
-        m => i += 'g'
-    ]);
+      ),
+      m => i += 'g'
+    );
 
     const stream = dispatch({});
     expect((await stream.next()).value).to.eql('a');
