+++ packages/mesh/src/sequence.ts
@@ -1,10 +1,10 @@
-import { FanoutDispatcherTargetsParamType, createFanoutDispatcher } from "./fanout";
+import { combine } from "./combine";
 
 /**
  * Executes a message against all target dispatchers in one after the other.
  */
 
-export const createSequenceDispatcher = <T>(targets: FanoutDispatcherTargetsParamType<T>) => createFanoutDispatcher(targets,  <T>(items:T[], each: (value: T) => Promise<any>) => {
+export const sequence = <T>(...fns: Function[]) => combine(fns,  <T>(items:T[], each: (value: T) => Promise<any>) => {
   return new Promise((resolve, reject) => {
     const next = (index: number) => {
       if (index === items.length) return resolve();
