+++ packages/mesh/src/wrap-promise.ts
@@ -1,0 +1,5 @@
+
+export function wrapPromise<TValue>(value: TValue | Promise<TValue>): Promise<TValue> {
+  if (value && value["then"]) return value as Promise<TValue>;
+  return Promise.resolve(value);
+}
